requirements.txt
# 车票生成API服务依赖
flask>=2.0.0
flask-cors>=3.0.0
pillow>=8.0.0
qrcode>=7.0.0
requests>=2.25.0

# 可选：用于更好的性能
gunicorn>=20.0.0
